137531 by jshimkus on 2018/12/14 (pair: sweettea) [VDO-4228], Merge 137454 to Phosphorus. 137454 by jshimkus on 2018/12/12 (pair: sweettea) [VDO-4228], Address dm target name conflict.
[BZ1588083]
vdo manager would incorrectly create a config file entry for a named
vdo even if the name was already in use as a dm target.  This would
result, at remove, in the destruction of the dm target.

This changeset prevents attempted creation of a vdo with a name that
conflicts with an extant dm target.  It also addresses cleanup
situations where the vdo config is out of sync with actuality and
has an old entry which has the same name as another non-vdo dm target.

.../VDOService.py
Modified to prevent creation of a vdo with name of an extant dm target.
Modified to handle cleanup scenarios with the same name conflict.

Co-authored-by: jshimkus <jshimkus@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>