137555 by tomj on 2018/12/17 (pair: sclafani) ALB-2991 Merge change 137475 to Homer 137475 by tomj on 2018/12/12 (pair: sclafani) ALB-2991 Record how an index gets opened
ALB-2991 is aiming to have the ability to suspend VDO while saving the
UDS index, followed by shutting down VDO without needing to do any more
writes to the disk. We will need to put the UDS index into a state
where we can close the index without saving the index to disk.

Loading an index can put us into a similar state. When we load a
cleanly saved index, we should be able to close the index without
saving what was just loaded.

This change records how the index was opened and will allow closing the
index to skip saving the index if the open was able to load a cleanly
saved index. Therefore we will always save an index upon closing
whenever the index was created new, or did a replay from a checkpoint,
or did a full index rebuild. When either udsLoadLocalIndex or
udsRebuildLocalIndex is able to cleanly load a saved index, we will
not need to do a save unless there have been some index queries.

.../index.h
.../index.c
.../indexInternals.c
Add the loadedType field to the Index. At index load time, record how
the index was loaded: LOAD_CREATE when it is a new index, LOAD_LOAD
when we loaded a saved index, LOAD_REPLAY when we did a replay from a
checkpoint, LOAD_EMPTY when we wanted to rebuild the master index but
found an empty volume, LOAD_REBUILD when we rebuilt the master index
from the volume, or LOAD_UNDEFINED when loading the index failed.

.../loadType.h
Add the value LOAD_EMPTY to identify an index that wanted to replay the
volume but found no valid chapters.
Add the value LOAD_REPLAY to identify an index that got loaded from a
checkpoint and needed some chapters to be replayed.
Document what the values in a LoadType really mean.

137560 by tomj on 2018/12/17 (pair: sclafani)
ALB-2991 Merge change 137497 to Homer
137497 by tomj on 2018/12/13 (pair: sclafani)
ALB-2991 Do not stop the chapter writer thread in the saveIndex method.

At the Index layer of the code, the major impediment to saving the
index while leaving the index open is that the saveIndex method stops
the chapter writer thread. This change modifies the saveIndex method
to not take any action that will stop or free the index.

We explicitly document the requirements that the caller of saveIndex
must meet before using the saveIndex method.

This change appears to remove the stopping of the chapter writer
thread. In fact there is also a call that stops this thread as
part of the freeChapterWriter method.

.../index.h
Document that it is now the explicit responsibility of caller to ensure
that there are no index requests in progress during the call to the
saveIndex method.

.../index.c
Remove the termination of the chapter writer thread from the saveIndex
method. Replace it with waiting for the chapter writer to be idle.
Log when we finish saving the index. This makes debugging the
testing of the ALB-2991 changes easier. It is also desirable
for understanding the VDO-4394 failure.

137566 by tomj on 2018/12/17 (pair: sclafani)
ALB-2991 Merge change 137537 to Homer
137537 by tomj on 2018/12/14 (pair: sclafani)
ALB-2991 Delete the saved open chapters of a recently saved index.

When the chapter writer writes its first chapter, it destroys the
saved open chapter. Now that we will be able to save an index without
freeing it, when we write the next chapter we also need to to destroy
the newly saved open chapter.

.../index.h
.../chapterWriter.c
Replace the firstChapterWrite flag with a new flag kept in the index
structure.

.../index.c
.../indexInternals.c
When we have made a new index, set the the hasSavedOpenChapter flag to
true if we cleanly loaded an existing saved index.
When we have saved an index, set the hasSavedOpenChapter flag.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sclafani <sclafani@redhat.com>