137587 by corwin on 2018/12/18 (pair: sweettea) Merged 137271 to linux-vdo. 137271 by corwin on 2018/11/30 (pair: sclafani) Improved notifyZoneStoppedScrubbing() [VDO-4402].
This bug entails adding a suspend method to all VDO components to
ensure that there is no I/O while the VDO is suspended. Before
implementing the suspend, some clean up of action handling in the slab
depot is in order. This is a small part of that clean up.

.../slabDepotInternals.h
.../slabDepot.c
Added result parameter to notifyZoneStoppedScrubbing() so that it can
report any errors from the scrubber. This probably doesn't matter in
practice since any such errors would have already put the VDO in
read-only mode. However, since the completion which is used to notify
the VDO when scrubbing is done in all zones has an error handler
registered, we might as well call it.
Replaced CAS loop in notifyZoneStoppedScrubbing() with a simple call
to atomicAdd32().

.../blockAllocator.c
Reflected changes to notifyZoneStoppedScrubbing().

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>