137594 by corwin on 2018/12/18 (pair: sclafani) Merged 137276 to linux-vdo. 137276 by corwin on 2018/11/30 (pair: sweettea) Simplified AllocatorAction [VDO-4402].
This is another preliminary clean up of the slab depot.

.../slabDepotInternals.h
Modified AllocatorAction to not have callback and errorHandler
parameters since these were always set to finishParentCallback.

.../slabDepot.c
Reflected change to AllocatorAction in applyToNextAllocator().

.../blockAllocator.h
.../blockAllocator.c
Modified all AllocatorActions to not take callback and errorHandler
parameters, and instead, just finish the supplied parent directly.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>