137616 by corwin on 2018/12/19 (pair: sweettea) Merged 137374 to linux-vdo 137374 by corwin on 2018/12/06 (pair: msakai, sweettea) Factored a generic ActionManager out of SlabDepot [VDO-4402].
This is the last of prequels before adding suspension of all I/O
when suspending the VDO. It creates a generic mechanism for applying
operations to multi-zone constructs such as the set of block allocators
or the set of block map zones.

.../actionManager.h
.../actionManager.c
.../types.h
Created a generic action manager based on the action methods from
slabDepot.c.

.../slabDepot.h
.../slabDepotInternals.h
.../slabDepot.c
Added a layer parameter to prepareToGrowSlabDepot() since the
completions which the SlabDepot used to get the layer from went away.
Added an ActionManager to SlabDepot and converted all actions to use
it.
Removed all of SlabDepot's own action machinery.

.../blockAllocator.h
.../blockAllocator.c
Converted AllocatorActions to ZoneActions.

.../vdoResize.c
Reflected change to prepareToGrowSlabDepot().

.../blockMapInternals.h
.../blockMap.c
Added an ActionManager to BlockMap and converted all actions to
use the ActionManager. This includes adding an aging flag to
prevent advanceBlockMapEra() from scheduling multiple pending
actions.
Removed all of BlockMap's own action machinery.

.../blockMap.c
.../forest.c
Factored all of the logic for installing forests or increasing the size
of the block map without changing the forest into replaceForest().

.../completion.h
.../completion.c
Added ACTION_COMPLETION which is used by ActionManager.
Removed BLOCK_MAP_COMPLETION and SLAB_DEPOT_COMPLETION since they are
no longer used.

.../recoveryJournal.h
.../recoveryJournal.c
Added getRecoveryJournalThreadID() for use by BlockMap when making its
ActionManager.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>