137626 by corwin on 2018/12/19 (pair: sclafani) Merged 137538 to linux-vdo. 137538 by corwin on 2018/12/14 (pair: sclafani) Removed vestiges of split-brain detection [VDO-4402].
Another small piece of reworking the read-only mode entering process
for suspend/resume.

.../threadData.h
.../threadData.c
Removed SuperBlockAccessState field and enum, and code which referenced
it as it was used to mediate super block access between the read-only
super block save, and the split brain detection super block read,
the latter of which, no longer exists.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>