137627 by corwin on 2018/12/19 (pair: sclafani) Merged 137572 to linux-vdo. 137572 by corwin on 2018/12/17 (pair: sclafani) Removed hash, logical, and physical zones use of ThreadData [VDO-4402].
This change is another small piece of the ThreadData clean up. By
eliminating on unnecessary dependency on the internals of ThreadData,
it becomes less onerous to privatize those internals.

.../hashZone.c
.../logicalZone.c
.../physicalZone.c
Replaced 'threadData' field with 'threadID" since it is only this
ID which is used by the zones. When the original parallelism story
was implemented back in Oxygen, it was thought that there would be
tighter integration between the ThreadData and the zones. This turns
out not to be the case.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>