137631 by corwin on 2018/12/19 (pair: sweettea) Merged 137588 and 137590 to linux-vdo. 137588 by corwin on 2018/12/18 (pair: sweettea) Made definition of ThreadData private [VDO-4402].
As part of the clean up of ThreadData, this change merely moves the
structure definition from threadData.h to a new threadData.c before
making changes to that structure so that diffs of those changes are
easier to read.

.../threadData.h
.../threadDataInternals.h
Moved declartion of struct threadData to new threadDataInternals.h.

.../threadData.c
.../vdo.c
Included new header file.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>