137834 by corwin on 2019/01/16 (pair: msakai) Merged 137753 to linux-vdo. 137753 by corwin on 2019/01/09 (pair: msakai, sweettea) Added re-enabling of read-only mode entry [VDO-4402].
This is the final clean up of ThreadData/ReadOnlyNotifier for the
suspend/resume work.

When waitUntilNotEnteringReadOnlyMode() is called on a
ReadOnlyNotifier, that method not only waits for any outstanding
read-only notifications to complete, but also puts the notifier in a
state which prevents subsequent calls to enterReadOnlyMode() from
sending notifications. Currently, this is not an issue, but once the
VDO suspend is modified to drain and prevent I/O until the VDO is
resumed, it will be necessary to re-enable notifications. This change
adds a new method on the notifier to do this. Furthermore, if some
operation encounters an unrecoverable error while the VDO is suspended,
the notifier must retain enough state to put the VDO into read-only
mode (and attempt to write the super block with the read-only state)
when it is resumed.

In addition to adding this new method, this change also simplifies the
notification mechanism. Hopefully this makes the implementation of the
new method, as well as reasoning about the entire notifier, easier to
do. The notifier has some fields common to the entire VDO, and also
some per-thread fields (the ThreadData structure). Prior to this
change, each ThreadData contained a completion. Calling
enterReadOnlyMode() from some base thread would use that thread's
completion to perform the notification by sending that completion to
the admin thread to check whether some notification has already begun.
If not, that completion would visit each of the threads to do the
notification before being sent back to its owning thread to finish. If
some other notification was already in progress, the completion would
be sent straight back to its owning thread. Calling
waitUntilNotEnteringReadOnlyMode() would also send a completion around
to each thread, checking whether that thread had an outstanding
notification, if so, when that notification finished, the waiter
completion would be sent to the next thread, otherwise, it would
proceed immediately. Since in practice, only one round of notification
is required, only one completion is actually necessary if there is
appropriate coordination among the threads. Furthermore, the rendezvous
required by the drain mechanism is tricky to extend to provide the
ability to delay read-only mode entry while the VDO is suspended, and
to perform it on resume.

With this change, the notifier has a single completion which is used to
do a read-only notification, and the ThreadData structure no longer
needs a completion at all. Furthermore, the additional completion which
the notifier used for draining notifications is also no longer needed.
Instead, the notifier uses a pair of atomics, one to hold the read-only
error and one to hold the current state of the notifier. The state is
used to control access to the completion and keep track of whether a
notification is in progress or needs to be performed.

.../readOnlyNotifier.h
.../readOnlyNotifier.c
Added allowReadOnlyModeEntry().
Added atomic 'readOnlyError' and 'state' fields to ReadOnlyNotifier.
Along with repurposing the notifier's completion for notifications
instead of draining, this facilitates removing all the ThreadData
fields other than 'isReadOnly' and the listeners list. Access to
the notifier's one and only completion is gated by the new atomic
fields. The ThreadData.superBlockIdleWaiter field is replaced by
a single 'waiter' field in the notifier since there can be at most
one such waiter at a time anyway. ThreadData's readOnlyModeWaiter
field was already unused.
Modified asNotifier() to reflect the repurposing of the notifier's
completion.
Removed asThreadData().
Modified waitUntilNotEnteringReadOnlyMode() to merely update the
state to MAY_NOT_NOTIFY if notifications are currently allowed, and
to wait for any in progress notification; removed
waitUntilThreadNotEnteringReadOnlyMode().
Modified enterReadOnlyMode() to use the state variable to decide
whether to start a notification or to record the error for a later
notification if notifications are not allowed.
Moved the read-only error logging to makeThreadReadOnly(). Removed
startNotification() because it is now superfluous;
enterReadOnlyMode() now calls makeThreadReadOnly() directly.
Modified finishEnteringReadOnlyMode() to use the new state
variables. This fixes a bug introduced in change 137720 which
runs this method on the admin thread instead of on the
initiating thread. With the current change, running this method
on the admin thread is now correct. Fortunately, this bug had no
visible effects.

.../completion.h
.../completion.c
Removed now unused WAIT_FOR_READ_ONLY_MODE_COMPLETION.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>