137931 by sclafani on 2019/01/30 (pair: msakai) Merged fixes for disabled statistics from //eng/vdo to linux-vdo 137892 by sclafani on 2019/01/25 (pair: msakai) Fixed bogus stat values when in read-only mode [BZ 1668747] [VDO-4429]
Stats that aren't defined in read-only mode should be rendered as
"N/A", but the read-only mode check was broken by the initial
python3 conversion way back in change 133133. That change made
string literals unicode by default, but the mode value from the
stat structure is a byte string, and the two can't ever be equal.

.../VDOStatistics.py
Converted mode string literals back to byte strings.

137916 by sclafani on 2019/01/29 (pair: jwiele)
Fixed display of disabled stats as zero on systems with python 3.6+

Testing of the fix for [BZ 1668747] [VDO-4429] revealed a
secondary issue: disabled stats would display as "0" instead of
"N/A" on systems with newer pythons (Fedora 28). We use a subclass
of int called NotAvailable for the disabled stats. Python has
changed somehow so that int(NotAvailable()) gets converted to an
int instead of staying NotAvailable despite overriding __int__().

.../Field.py
The call to int() currently triggering the problem is in IntField,
so worked around the issue by making IntField check if the value
is already an int instance, skipping the int() call if it is.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>