138263 by corwin on 2019/03/12 (pair: msakai) Merged 137799 to linux-vdo. 137799 by corwin on 2019/01/14 (pair: msakai) Removed ADMIN_STATE_CLOSE_REQUESTED [VDO-4402].
This is a small prequel to implementing suspend/resume. Prior to this
change, ADMIN_STATE_CLOSE_REQUESTED is used to capture the state in the
block map trees where a close has been requested, but lookups are still
active. In practice, this state never exists because the close
transaction (which will become the suspend/save transaction) is never
initiated until after there are no active DataVIOs.

.../adminState.h
Removed ADMIN_STATE_CLOSE_REQUESTED and modified isClosing() to reflect
the removal.

.../blockMapTree.c
Made it an error to call closeZoneTrees() if there are active lookups
since in practice we never do this.
Removed check in finishedLookup() to see if the tree can close since
close may not be (and isn't) initiated before all lookups are
complete.

.../statusCodes.h
.../statusCodes.c
Added VDO_INVALID_ADMIN_STATE for the new error check in
closeZoneTrees(). It will see much more use in the coming
suspend/resume implementation.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>