138347 by bjohnston on 2019/03/18 (pair: msakai) [VDO-4447] Merge 138139 and 138156 to vdo linux. 138139 by bjohnston on 2019/02/28 (pair: msakai) [BZ 1684249] [VDO-4447] Set scan_lvs on in lvm.conf for pvcreate.
.../VDOService.py
Call pvcreate with the --config option to set the lvm.conf
property scan_lvs on. This will allow pvcreate --test to
work again.

138156 by bjohnston on 2019/03/04 (pair: sweettea)
[BZ 1684249] [VDO-4447] Fix error parsing for pvcreate.

.../VDOService.py
Currently when pvcreate fails we parse the stderr for a
specific error message at a specific line number. With
the change to pvcreate to add the --config option, an
additional line can now show up in stderr when run
on a machine where the lvm patch has not yet been
applied. Modify the parsing code to be more generic
about how many lines are in stderr.

Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>