138387 by jwiele on 2019/03/19 (pair: sweettea) VDO-4446 Merge change 138355 from //eng/vdo to Phosphorus. 138355 by jwiele on 2019/03/18 (pair: joe,sweettea) VDO-4446 Check logfile and confFile arguments more carefully. BZ1688481 for RHEL7.7, BZ1679224 for RHEL8.1.
With this change, any invalid arguments to the checkConfFile
and checkLogfile arguments should be rejected with an error
message instead of a Python stack trace.

.../Defaults.py
Expanded checkLogFile to reject directories in addtion to
block devices as arguments to the logfile options with
reasonable error messages. As a bonus, due to shared code,
checkConfFile gets better checking as well.

.../vdo
Changed how a handler is created for a logfile path passed as
a command line option. If the path is an existing file or
does not exist yet, a RotatingFile handler is attempted. If
the path exists but is not a file, a StreamHandler is
attempted. Redirecting log output to a non-file device,
e. g. /dev/stderr redirected to a pipe, has always been
supported as long as the device supports write
semantics. According to e. g.,
https://docs.python.org/release/3.4.0/library/logging.handlers.html#streamhandler,
StreamHandler is the correct handler for such a device.

Block devices and directories will have been filtered out
during argument parsing with a descriptive error message
(see Defaults.py). Any error raised while setting up the
handler will result in a generic error message to the effect
that logging cannot be configured using the path. Examples
of such errors are: lack of write permissions (e. g., an
NFS-mounted path with root squash), the path representing a
device node that cannot be written, or the path having been
changed asynchronously by some other program between
argument parsing and handler setup.

Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>